generate serialized KLL sketches for C++
